LifeHelper

+++ app/src/main/java/com/ns/yc/lifehelper/ui/other/gank/presenter/GanKWebPresenter.java
@@ -26,6 +26,7 @@ public class GanKWebPresenter implements GanKWebContract.Presenter {
     private boolean mIsFavorite;
     private GanKFavorite favoriteData;
     private Realm realm;
+    private RealmResults<GanKFavorite> ganKFavorites;
 
 
     public GanKWebPresenter(GanKWebContract.View view) {
@@ -57,7 +58,9 @@ public class GanKWebPresenter implements GanKWebContract.Presenter {
     }
 
     private void initRealm() {
-        realm = BaseApplication.getInstance().getRealmHelper();
+        if(realm==null){
+            realm = BaseApplication.getInstance().getRealmHelper();
+        }
     }
 
 
@@ -65,6 +68,7 @@ public class GanKWebPresenter implements GanKWebContract.Presenter {
         mUrl = mView.getLoadUrl();
         mView.loadURL(mUrl);
     }
+
 
     /**
      * 处理收藏逻辑
@@ -78,12 +82,38 @@ public class GanKWebPresenter implements GanKWebContract.Presenter {
         }
     }
 
+
     /**
      * 收藏
      */
     private void favorite() {
         favoriteData.setCreatetime(System.currentTimeMillis());
-        mIsFavorite = favoriteData.isManaged();
+        initRealm();
+        if(realm!=null && realm.where(GanKFavorite.class)!=null){
+            ganKFavorites = realm.where(GanKFavorite.class).findAll();
+        }else {
+            ToastUtils.showShort("收藏失败,请重试");
+            return;
+        }
+        realm.beginTransaction();
+        //RealmException: 'GanKFavorite' has a primary key, use 'createObject(Class<E>, Object)' instead.
+        //GanKFavorite favorite = realm.createObject(GanKFavorite.class);
+        //Primary key field 'gankID' cannot be changed after object was created.
+        //GanKFavorite favorite = realm.createObject(GanKFavorite.class , UUID.randomUUID().toString());
+        //GanKFavorite favorite = realm.createObject(GanKFavorite.class,"gankID");
+        GanKFavorite favorite = realm.createObject(GanKFavorite.class);
+        favorite.setTitle(favoriteData.getTitle());
+        favorite.setUrl(favoriteData.getUrl());
+        favorite.setType(favoriteData.getType());
+        favorite.setGankID(favoriteData.getGankID());
+        favorite.setAuthor(favoriteData.getAuthor());
+        favorite.setCreatetime(favoriteData.getCreatetime());
+        favorite.setData(favoriteData.getData());
+        realm.copyToRealm(favorite);
+        realm.commitTransaction();
+        //mIsFavorite = favoriteData.isManaged();
+        //判断数据库是否加载数据成功
+        mIsFavorite = favoriteData.isLoaded();
         mView.setFavoriteState(mIsFavorite);
         if (!mIsFavorite) {
             ToastUtils.showShort("收藏失败,请重试");
@@ -92,23 +122,30 @@ public class GanKWebPresenter implements GanKWebContract.Presenter {
         }
     }
 
+
     /**
      * 取消收藏
      */
     private void unFavorite() {
-        if(realm!=null){
+        initRealm();
+        if(realm!=null && realm.where(GanKFavorite.class).findAll()!=null){
             mIsFavorite = realm.where(GanKFavorite.class)
                     .equalTo("gankID", favoriteData.getGankID())
                     .findAll()
                     .deleteAllFromRealm();
+        }else {
+            ToastUtils.showShort("取消收藏失败,请重试");
+            return;
         }
         // 不调用这句保存 在保存会失败，并且返回的是true
         favoriteData.deleteFromRealm();
+        mIsFavorite = favoriteData.isLoaded();
         mView.setFavoriteState(mIsFavorite);
         if (mIsFavorite) {
             ToastUtils.showShort("取消收藏失败,请重试");
         }
     }
+
 
     /**
      * 是否有收藏
@@ -119,6 +156,7 @@ public class GanKWebPresenter implements GanKWebContract.Presenter {
             mView.hideFavoriteFab();
             return;
         }
+        initRealm();
         if(realm!=null && realm.where(GanKFavorite.class)!=null){
             RealmResults<GanKFavorite> favorites = realm.where(GanKFavorite.class)
                     .equalTo("gankID", favoriteData.getGankID())
